Added multiple access support for rtimer
the rtimer module can now be compiled with RTIMER_MULTIPLE_ACCESS
enabled so that multiple rtimers can be scheduled at once.

For this to work implementations of atomic_section_enter and
atomic_section_exit are required.

In addition the rtimer_cancel function has been provided which
also works with RTIMER_MULTIPLE_ACCESS == 0